Fuse: Split regression tests (without/with a server runtime) #425
