+++ tests/org.jboss.tools.fuse.ui.bot.test/src/org/jboss/tools/fuse/ui/bot/test/RegressionTest.java
@@ -18,15 +18,10 @@ import javax.xml.parsers.ParserConfigurationException;
 import org.jboss.reddeer.eclipse.condition.ConsoleHasText;
 import org.jboss.reddeer.eclipse.jdt.ui.ProjectExplorer;
 import org.jboss.reddeer.eclipse.ui.console.ConsoleView;
-import org.jboss.reddeer.junit.requirement.inject.InjectRequirement;
 import org.jboss.reddeer.junit.runner.RedDeerSuite;
-import org.jboss.reddeer.requirements.cleanworkspace.CleanWorkspaceRequirement.CleanWorkspace;
 import org.jboss.reddeer.requirements.openperspective.OpenPerspectiveRequirement.OpenPerspective;
-import org.jboss.reddeer.requirements.server.ServerReqState;
-import org.jboss.reddeer.swt.api.TreeItem;
 import org.jboss.reddeer.swt.condition.ShellWithTextIsAvailable;
 import org.jboss.reddeer.swt.exception.SWTLayerException;
-import org.jboss.reddeer.swt.impl.button.CheckBox;
 import org.jboss.reddeer.swt.impl.button.PushButton;
 import org.jboss.reddeer.swt.impl.ctab.DefaultCTabItem;
 import org.jboss.reddeer.swt.impl.menu.ContextMenu;
@@ -41,27 +36,18 @@ import org.jboss.reddeer.swt.matcher.WithTooltipTextMatcher;
 import org.jboss.reddeer.swt.wait.AbstractWait;
 import org.jboss.reddeer.swt.wait.TimePeriod;
 import org.jboss.reddeer.swt.wait.WaitUntil;
-import org.jboss.reddeer.swt.wait.WaitWhile;
 import org.jboss.reddeer.workbench.impl.editor.DefaultEditor;
 import org.jboss.tools.fuse.reddeer.component.Log;
-import org.jboss.tools.fuse.reddeer.condition.FuseLogContainsText;
-import org.jboss.tools.fuse.reddeer.debug.BreakpointsView;
-import org.jboss.tools.fuse.reddeer.debug.IsRunning;
 import org.jboss.tools.fuse.reddeer.debug.IsSuspended;
-import org.jboss.tools.fuse.reddeer.debug.ResumeButton;
 import org.jboss.tools.fuse.reddeer.editor.CamelEditor;
 import org.jboss.tools.fuse.reddeer.perspectives.FuseIntegrationPerspective;
 import org.jboss.tools.fuse.reddeer.preference.ServerRuntimePreferencePage;
 import org.jboss.tools.fuse.reddeer.projectexplorer.CamelProject;
-import org.jboss.tools.fuse.reddeer.server.ServerManipulator;
 import org.jboss.tools.fuse.reddeer.utils.ResourceHelper;
 import org.jboss.tools.fuse.reddeer.view.JMXNavigator;
 import org.jboss.tools.fuse.ui.bot.test.utils.EditorManipulator;
 import org.jboss.tools.fuse.ui.bot.test.utils.FuseArchetypeNotFoundException;
 import org.jboss.tools.fuse.ui.bot.test.utils.ProjectFactory;
-import org.jboss.tools.runtime.reddeer.requirement.ServerReqType;
-import org.jboss.tools.runtime.reddeer.requirement.ServerRequirement;
-import org.jboss.tools.runtime.reddeer.requirement.ServerRequirement.Server;
 import org.junit.After;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -75,22 +61,13 @@ import org.xml.sax.SAXException;
  * 
  * @author tsedmik
  */
-@CleanWorkspace
 @OpenPerspective(FuseIntegrationPerspective.class)
 @RunWith(RedDeerSuite.class)
-@Server(type = ServerReqType.Fuse, state = ServerReqState.PRESENT)
 public class RegressionTest extends DefaultTest {
-
-	@InjectRequirement
-	private ServerRequirement serverRequirement;
 
 	@After
 	public void clean() {
 
-		String server = serverRequirement.getConfig().getName();
-		if (ServerManipulator.isServerStarted(server)) {
-			ServerManipulator.stopServer(server);
-		}
 		new ProjectExplorer().deleteAllProjects();
 	}
 
@@ -99,7 +76,6 @@ public class RegressionTest extends DefaultTest {
 	 * https://issues.jboss.org/browse/FUSETOOLS-674
 	 * 
 	 * NOTE: not fixed yet - deferred to 8.0
-	 * @throws FuseArchetypeNotFoundException 
 	 */
 	@Ignore
 	@Test
@@ -146,58 +122,6 @@ public class RegressionTest extends DefaultTest {
 			new ContextMenu("Debug As", "3 Local Camel Context (without tests)");
 		} catch (Exception e) {
 			fail("Context menu 'Debug As --> Local Camel Context' is missing");
-		}
-	}
-
-	/**
-	 * New Server Runtime Wizard - Cancel/Finish button error
-	 * https://issues.jboss.org/browse/FUSETOOLS-1067
-	 */
-	@Test
-	public void issue_1067() {
-
-		new ServerRuntimePreferencePage().open();
-
-		new PushButton("Add...").click();
-		new DefaultShell("New Server Runtime Environment").setFocus();
-		new DefaultTreeItem("JBoss Fuse").expand();
-
-		// tests the _Finish_ button
-		for (TreeItem item : new DefaultTreeItem("JBoss Fuse").getItems()) {
-			if (!item.getText().startsWith("JBoss"))
-				continue;
-			AbstractWait.sleep(TimePeriod.SHORT);
-			item.select();
-			try {
-
-				assertFalse(new PushButton("Finish").isEnabled());
-			} catch (AssertionError ex) {
-
-				new DefaultTreeItem("JBoss Fuse").select();
-				AbstractWait.sleep(TimePeriod.SHORT);
-				new PushButton("Cancel").click();
-				AbstractWait.sleep(TimePeriod.NORMAL);
-				new DefaultShell().close();
-				throw ex;
-			}
-		}
-
-		// tests the _Cancel_ button
-		AbstractWait.sleep(TimePeriod.SHORT);
-		new DefaultTreeItem("JBoss Fuse", serverRequirement.getConfig().getServerBase().getName()).select();
-		new PushButton("Cancel").click();
-		try {
-
-			assertTrue(new DefaultShell().getText().equals("Preferences"));
-		} catch (AssertionError ex) {
-
-			new DefaultShell().close();
-			new DefaultTreeItem("JBoss Fuse").select();
-			AbstractWait.sleep(TimePeriod.SHORT);
-			new PushButton("Cancel").click();
-			AbstractWait.sleep(TimePeriod.NORMAL);
-			new DefaultShell().close();
-			throw ex;
 		}
 	}
 
@@ -301,40 +225,6 @@ public class RegressionTest extends DefaultTest {
 	}
 
 	/**
-	 * Karaf cannot be started in debug mode
-	 * https://issues.jboss.org/browse/FUSETOOLS-1132
-	 * @throws FuseArchetypeNotFoundException 
-	 */
-	@Test
-	public void issue_1132() throws FuseArchetypeNotFoundException {
-
-		ProjectFactory.createProject("camel-blueprint", "camel-archetype-blueprint");
-		new ProjectExplorer().getProject("camel-blueprint").getProjectItem("src/test/java", "com.mycompany.camel.blueprint", "RouteTest.java").delete();
-		String server = serverRequirement.getConfig().getName();
-		new BreakpointsView().importBreakpoints(ResourceHelper.getResourceAbsolutePath(Activator.PLUGIN_ID, "resources/breakpoint.bkpt"));
-		ServerManipulator.debugServer(server);
-		ServerManipulator.addModule(server, "camel-blueprint");
-		try {	
-			new WaitUntil(new ShellWithTextIsAvailable("Confirm Perspective Switch"), TimePeriod.LONG);
-			new DefaultShell("Confirm Perspective Switch");
-			new CheckBox(0).toggle(true);
-			new PushButton("No").click();
-			AbstractWait.sleep(TimePeriod.NORMAL);
-		} catch (Exception e) {
-			fail();
-		} finally {
-			new WaitWhile(new IsRunning());
-			new BreakpointsView().removeAllBreakpoints();
-			new WorkbenchShell().setFocus();
-			ResumeButton resume = new ResumeButton();
-			if (resume.isEnabled()) {
-				resume.select();
-			}
-			ServerManipulator.removeAllModules(server);
-		}
-	}
-
-	/**
 	 * New Fuse Project - Finish button
 	 * https://issues.jboss.org/browse/FUSETOOLS-1149
 	 */
@@ -347,23 +237,6 @@ public class RegressionTest extends DefaultTest {
 			new DefaultShell().close();
 			fail("'Finish' button should not be enabled!");
 		}
-	}
-
-	/**
-	 * uninstall of bundles from servers broken
-	 * https://issues.jboss.org/browse/FUSETOOLS-1152
-	 * @throws FuseArchetypeNotFoundException 
-	 */
-	@Test
-	public void issue_1152() throws FuseArchetypeNotFoundException {
-
-		ProjectFactory.createProject("camel-spring-dm", "camel-archetype-spring-dm");
-		String server = serverRequirement.getConfig().getName();
-		ServerManipulator.startServer(server);
-		ServerManipulator.addModule(server, "camel-spring-dm");
-		AbstractWait.sleep(TimePeriod.NORMAL);
-		ServerManipulator.removeAllModules(server);
-		new WaitUntil(new FuseLogContainsText("Application context succesfully closed (OsgiBundleXmlApplicationContext(bundle=camel-spring-dm"), TimePeriod.VERY_LONG);
 	}
 
 	/**
@@ -462,29 +335,5 @@ public class RegressionTest extends DefaultTest {
 			new DefaultTreeItem("JBoss Fuse Launcher").select();
 			fail("Run Configurations contains forbidden item");
 		} catch (SWTLayerException e) {}
-	}
-
-	/**
-	 * Problem occurred during restart JBoss Fuse
-	 * https://issues.jboss.org/browse/FUSETOOLS-1252
-	 * @throws FuseArchetypeNotFoundException 
-	 */
-	@Test
-	public void issue_1252() throws FuseArchetypeNotFoundException {
-
-		ProjectFactory.createProject("camel-blueprint", "camel-archetype-blueprint");
-		String server = serverRequirement.getConfig().getName();
-		ServerManipulator.addModule(server, "camel-blueprint");
-		ServerManipulator.startServer(server);
-		ServerManipulator.restartInDebug(server);
-		try {
-			new WaitUntil(new ShellWithTextIsAvailable("Problem Occurred"));
-			new DefaultShell("Problem Occurred");
-			new PushButton("OK");
-		} catch (Exception e) {
-			// OK no shell "Problem Occurred" was found
-			return;
-		}
-		fail();
 	}
 }
